thymeleaf이용해서 같은 form 합치기

+++ SpringBoot_Mybatis_Thymeleaf/src/main/java/com/iu/s1/board/BoardMapper.java
@@ -15,9 +15,9 @@ public interface BoardMapper {
 	//Insert
 	public Long setBoard(BoardVO boardVO)throws Exception;
 	//Update
-	public Integer setUpdate(BoardVO boardVO)throws Exception;
+	public Long setUpdate(BoardVO boardVO)throws Exception;
 	//HitUpdate
-	public Integer setHitUpdate(BoardVO boardVO)throws Exception;
+	public Long setHitUpdate(BoardVO boardVO)throws Exception;
 	//Delete
-	public Integer setDelete(BoardVO boardVO)throws Exception;
+	public Long setDelete(BoardVO boardVO)throws Exception;
 }
